feat:完成下拉刷新&&上拉加载的封装

+++ base/lib/entity/circle/circle_dynamic_bean.dart
@@ -101,8 +101,8 @@ class CircleLinkBean {
 @JsonSerializable()
 class CircleVideoBean {
   String url;
-  double width;
-  double height;
+  int width;
+  int height;
 
   CircleVideoBean({this.url, this.width, this.height});
